+++ doc/source/conf.py
@@ -125,8 +125,9 @@ man_pages = [
 
 intersphinx_mapping = {
     'python': ('https://docs.python.org/3/', None),
-    'requests': ('http://docs.python-requests.org/en/master/', None),
     'ClusterShell': ('http://clustershell.readthedocs.io/en/v1.8/', None),
+    'cumin': ('https://doc.wikimedia.org/cumin/master/', None),
+    'requests': ('http://docs.python-requests.org/en/master/', None),
 }
 
 # Napoleon settings
@@ -144,7 +145,8 @@ napoleon_use_rtype = True
 napoleon_use_keyword = True
 
 # Autodoc settings
-autodoc_default_flags = ['members', 'show-inheritance', 'inherited-members', 'private-members']
+autodoc_default_flags = ['members', 'show-inheritance', 'inherited-members']
+autoclass_content = 'both'
 
 
 # -- Helper functions -----------------------------------------------------