+++ spicerack/administrative.py
@@ -20,6 +20,10 @@ class Reason:
             username (str): the username to mention in the reason as the author of the action.
             hostname (str): the hostname to mention in the reason as the host originating the action.
             task_id (str, optional): the task ID to mention in the reason.
+
+        Raises:
+            spicerack.administrative.ReasonError: if any parameter contains double quotes.
+
         """
         self._reason = reason
         self._username = username
@@ -48,14 +52,21 @@ class Reason:
             str: the generated string representation of all the instance attributes.
 
         """
-        parts = [
-            self._reason,
-            '{user}@{host}'.format(user=self._username, host=self._hostname),
-        ]
+        parts = [self._reason, self.owner]
         if self._task_id is not None:
             parts.append(self._task_id)
 
         return ' - '.join(parts)
+
+    @property
+    def owner(self):
+        """Getter for the owner property.
+
+        Returns:
+            str: the origin (user@host) of the currently running code.
+
+        """
+        return '{user}@{host}'.format(user=self._username, host=self._hostname)
 
     def quoted(self):
         """Quoted string representation of the instance, including all attributes.