+++ spicerack/constants.py
@@ -1,4 +1,13 @@
 """Constants module."""
 
+ALL_DATACENTERS = ('eqiad', 'codfw', 'esams', 'ulsfo', 'eqsin')
+"""tuple: list of all datacenters."""
+
 CORE_DATACENTERS = ('eqiad', 'codfw')
 """tuple: list of core datacenters."""
+
+INTERNAL_TLD = 'wmnet'
+"""str: the name of the internal top level domain."""
+
+MANAGEMENT_SUBDOMAIN = 'mgmt'
+"""str: the name of the subdomain for the management interfaces."""