+++ spicerack/dns.py
@@ -22,14 +22,12 @@ class DnsNotFound(DnsError):
 class Dns:
     """Class to interact with the DNS."""
 
-    def __init__(self, *, nameserver_address=None, dry_run=True):
+    def __init__(self, *, nameserver_address=None):
         """Initialize the instance.
 
         Arguments:
             nameserver_address (str, optional): the nameserver address to use, if not set uses the OS configuration.
-            dry_run (bool, optional): whether this is a DRY-RUN.
         """
-        self._dry_run = dry_run
         if nameserver_address:
             self._resolver = dns.resolver.Resolver(configure=False)
             self._resolver.nameservers = [nameserver_address]
@@ -132,7 +130,7 @@ class Dns:
         """
         try:
             response = self._resolver.query(qname, record_type)
-            logger.debug('Resolved %s record for %s: %s', record_type, qname, response)
+            logger.debug('Resolved %s record for %s: %s', record_type, qname, response.rrset)
         except (dns.resolver.NoAnswer, dns.resolver.NXDOMAIN) as e:
             raise DnsNotFound('Record {record_type} not found for {qname}'.format(
                 record_type=record_type, qname=qname)) from e